# shop-ui/src/openapi/openapi.json
{
  "openapi": "3.1.0",
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/category-templates": {
      "get": {
        "tags": [
          "category-template-controller"
        ],
        "operationId": "getCategoryTemplates",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryTemplateResponseDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "category-template-controller"
        ],
        "operationId": "createCategoryTemplate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryTemplateRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryTemplateResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/category-templates/{category-template-id}": {
      "get": {
        "tags": [
          "category-template-controller"
        ],
        "operationId": "getCategoryTemplateById",
        "parameters": [
          {
            "name": "category-template-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryTemplateResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/category-templates/category/{category-id}": {
      "get": {
        "tags": [
          "category-template-controller"
        ],
        "operationId": "getCategoryTemplateByCategoryId",
        "parameters": [
          {
            "name": "category-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryTemplateResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/category-templates/{category-id}": {
      "delete": {
        "tags": [
          "category-template-controller"
        ],
        "operationId": "deleteCategoryTemplateByCategoryId",
        "parameters": [
          {
            "name": "category-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "PropertyPresetDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "value": {
            "type": "string"
          }
        }
      },
      "PropertyResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "presets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyPresetDTO"
            }
          }
        }
      },
...
      "CategoryTemplateRequestDTO": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "integer",
            "format": "int32"
          },
          "propertyIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          }
        },
        "required": [
          "categoryId",
          "propertyIds"
        ]
      },
      "CategoryTemplateResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "categoryId": {
            "type": "integer",
            "format": "int32"
          },
          "properties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyResponseDTO"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}